CIAM-1178 CLI: Support resources in iam commands - Add resources to list, create delete ccloud iam command - Unit Test

+++ internal/cmd/iam/command_rolebinding.go
@@ -3,13 +3,13 @@ package iam
 import (
 	"context"
 	"fmt"
+	"github.com/confluentinc/go-printer"
 	"net/http"
 	"sort"
 	"strings"
 
 	orgv1 "github.com/confluentinc/cc-structs/kafka/org/v1"
 
-	"github.com/confluentinc/go-printer"
 	mds "github.com/confluentinc/mds-sdk-go/mdsv1"
 	"github.com/confluentinc/mds-sdk-go/mdsv2alpha1"
 	"github.com/spf13/cobra"
@@ -51,13 +51,14 @@ var (
 )
 
 type rolebindingOptions struct {
-	role             string
-	resource         string
-	prefix           bool
-	principal        string
-	scopeV2          mdsv2alpha1.Scope
-	mdsScope         mds.MdsScope
-	resourcesRequest mds.ResourcesRequest
+	role               string
+	resource           string
+	prefix             bool
+	principal          string
+	scopeV2            mdsv2alpha1.Scope
+	mdsScope           mds.MdsScope
+	resourcesRequest   mds.ResourcesRequest
+	resourcesRequestV2 mdsv2alpha1.ResourcesRequest
 }
 
 type rolebindingCommand struct {
@@ -154,20 +155,20 @@ func (c *rolebindingCommand) init() {
 		RunE:    cmd.NewCLIRunE(c.list),
 		Example: example,
 	}
-	listCmd.Flags().String("principal", "", "Principal whose rolebindings should be listed.")
-	listCmd.Flags().Bool("current-user", false, "Show rolebindings belonging to current user.")
-	listCmd.Flags().String("role", "", "List rolebindings under a specific role given to a principal. Or if no principal is specified, list principals with the role.")
+	listCmd.Flags().String("principal", "", "Principal whose role bindings should be listed.")
+	listCmd.Flags().Bool("current-user", false, "Show role bindings belonging to current user.")
+	listCmd.Flags().String("role", "", "List role bindings under a specific role given to a principal. Or if no principal is specified, list principals with the role.")
+	listCmd.Flags().String("resource", "", "If specified with a role and no principals, list principals with role bindings to the role for this qualified resource.")
 	if c.cliName == "ccloud" {
-		listCmd.Flags().String("cloud-cluster", "", "Cloud cluster ID for scope of rolebinding listings.")
-		listCmd.Flags().String("environment", "", "Environment ID for scope of rolebinding listings.")
+		listCmd.Flags().String("cloud-cluster", "", "Cloud cluster ID for scope of role binding listings.")
+		listCmd.Flags().String("environment", "", "Environment ID for scope of role binding listings.")
 		listCmd.Flags().Bool("current-env", false, "Use current environment ID for scope.")
 	} else {
-		listCmd.Flags().String("kafka-cluster-id", "", "Kafka cluster ID for scope of rolebinding listings.")
-		listCmd.Flags().String("resource", "", "If specified with a role and no principals, list principals with rolebindings to the role for this qualified resource.")
-		listCmd.Flags().String("schema-registry-cluster-id", "", "Schema Registry cluster ID for scope of rolebinding listings.")
-		listCmd.Flags().String("ksql-cluster-id", "", "ksqlDB cluster ID for scope of rolebinding listings.")
-		listCmd.Flags().String("connect-cluster-id", "", "Kafka Connect cluster ID for scope of rolebinding listings.")
-		listCmd.Flags().String("cluster-name", "", "Cluster name to uniquely identify the cluster for rolebinding listings.")
+		listCmd.Flags().String("kafka-cluster-id", "", "Kafka cluster ID for scope of role binding listings.")
+		listCmd.Flags().String("schema-registry-cluster-id", "", "Schema Registry cluster ID for scope of role binding listings.")
+		listCmd.Flags().String("ksql-cluster-id", "", "ksqlDB cluster ID for scope of role binding listings.")
+		listCmd.Flags().String("connect-cluster-id", "", "Kafka Connect cluster ID for scope of role binding listings.")
+		listCmd.Flags().String("cluster-name", "", "Cluster name to uniquely identify the cluster for role binding listings.")
 	}
 	listCmd.Flags().StringP(output.FlagName, output.ShortHandFlag, output.DefaultValue, output.Usage)
 	listCmd.Flags().SortFlags = false
@@ -188,13 +189,13 @@ func (c *rolebindingCommand) init() {
 	}
 	createCmd.Flags().String("role", "", "Role name of the new role binding.")
 	createCmd.Flags().String("principal", "", "Qualified principal name for the role binding.")
+	createCmd.Flags().Bool("prefix", false, "Whether the provided resource name is treated as a prefix pattern.")
+	createCmd.Flags().String("resource", "", "Qualified resource name for the role binding.")
 	if c.cliName == "ccloud" {
 		createCmd.Flags().String("cloud-cluster", "", "Cloud cluster ID for the role binding.")
 		createCmd.Flags().String("environment", "", "Environment ID for scope of rolebinding create.")
 		createCmd.Flags().Bool("current-env", false, "Use current environment ID for scope.")
 	} else {
-		createCmd.Flags().Bool("prefix", false, "Whether the provided resource name is treated as a prefix pattern.")
-		createCmd.Flags().String("resource", "", "Qualified resource name for the role binding.")
 		createCmd.Flags().String("kafka-cluster-id", "", "Kafka cluster ID for the role binding.")
 		createCmd.Flags().String("schema-registry-cluster-id", "", "Schema Registry cluster ID for the role binding.")
 		createCmd.Flags().String("ksql-cluster-id", "", "ksqlDB cluster ID for the role binding.")
@@ -215,13 +216,13 @@ func (c *rolebindingCommand) init() {
 	}
 	deleteCmd.Flags().String("role", "", "Role name of the existing role binding.")
 	deleteCmd.Flags().String("principal", "", "Qualified principal name associated with the role binding.")
+	deleteCmd.Flags().Bool("prefix", false, "Whether the provided resource name is treated as a prefix pattern.")
+	deleteCmd.Flags().String("resource", "", "Qualified resource name for the role binding.")
 	if c.cliName == "ccloud" {
 		deleteCmd.Flags().String("cloud-cluster", "", "Cloud cluster ID for the role binding.")
 		deleteCmd.Flags().String("environment", "", "Environment ID for scope of rolebinding delete.")
 		deleteCmd.Flags().Bool("current-env", false, "Use current environment ID for scope.")
 	} else {
-		deleteCmd.Flags().Bool("prefix", false, "Whether the provided resource name is treated as a prefix pattern.")
-		deleteCmd.Flags().String("resource", "", "Qualified resource name associated with the role binding.")
 		deleteCmd.Flags().String("kafka-cluster-id", "", "Kafka cluster ID for the role binding.")
 		deleteCmd.Flags().String("schema-registry-cluster-id", "", "Schema Registry cluster ID for the role binding.")
 		deleteCmd.Flags().String("ksql-cluster-id", "", "ksqlDB cluster ID for the role binding.")
@@ -262,7 +263,25 @@ func (c *rolebindingCommand) parseAndValidateResourcePattern(typename string, pr
 	return result, nil
 }
 
-func (c *rolebindingCommand) validateRoleAndResourceType(roleName string, resourceType string) error {
+func (c *rolebindingCommand) parseAndValidateResourcePatternV2(typename string, prefix bool) (mdsv2alpha1.ResourcePattern, error) {
+	var result mdsv2alpha1.ResourcePattern
+	if prefix {
+		result.PatternType = "PREFIXED"
+	} else {
+		result.PatternType = "LITERAL"
+	}
+
+	parts := strings.Split(typename, ":")
+	if len(parts) != 2 {
+		return result, errors.NewErrorWithSuggestions(errors.ResourceFormatErrorMsg, errors.ResourceFormatSuggestions)
+	}
+	result.ResourceType = parts[0]
+	result.Name = parts[1]
+
+	return result, nil
+}
+
+func (c *rolebindingCommand) validateRoleAndResourceTypeV1(roleName string, resourceType string) error {
 	ctx := c.createContext()
 	role, resp, err := c.MDSClient.RBACRoleDefinitionsApi.RoleDetail(ctx, roleName)
 	if err != nil || resp.StatusCode == 204 {
@@ -320,6 +339,31 @@ func (c *rolebindingCommand) validateResourceType(resourceType string) error {
 	}
 
 	return nil
+}
+
+func (c *rolebindingCommand) validateRoleAndResourceTypeV2(roleName string, resourceType string) error {
+	ctx := c.createContext()
+	role, resp, err := c.MDSv2Client.RBACRoleDefinitionsApi.RoleDetail(ctx, roleName, nil)
+	if err != nil || resp.StatusCode == http.StatusNoContent {
+		if err == nil {
+			return errors.NewErrorWithSuggestions(fmt.Sprintf(errors.LookUpRoleErrorMsg, roleName), errors.LookUpRoleSuggestions)
+		} else {
+			return errors.NewWrapErrorWithSuggestions(err, fmt.Sprintf(errors.LookUpRoleErrorMsg, roleName), errors.LookUpRoleSuggestions)
+		}
+	}
+
+	var allResourceTypes []string
+	for _, policies := range role.Policies {
+		for _, operation := range policies.AllowedOperations {
+			allResourceTypes = append(allResourceTypes, operation.ResourceType)
+			if operation.ResourceType == resourceType {
+				return nil
+			}
+		}
+	}
+
+	suggestionsMsg := fmt.Sprintf(errors.InvalidResourceTypeSuggestions, strings.Join(allResourceTypes, ", "))
+	return errors.NewErrorWithSuggestions(fmt.Sprintf(errors.InvalidResourceTypeErrorMsg, resourceType), suggestionsMsg)
 }
 
 func (c *rolebindingCommand) parseAndValidateScope(cmd *cobra.Command) (*mds.MdsScope, error) {
@@ -674,7 +718,7 @@ func (c *rolebindingCommand) confluentListRolePrincipals(cmd *cobra.Command, opt
 		if err != nil {
 			return err
 		}
-		err = c.validateRoleAndResourceType(role, resource.ResourceType)
+		err = c.validateRoleAndResourceTypeV1(role, resource.ResourceType)
 		if err != nil {
 			return err
 		}
@@ -717,12 +761,38 @@ func (c *rolebindingCommand) ccloudListRolePrincipals(cmd *cobra.Command, option
 	scopeV2 := &options.scopeV2
 	role := options.role
 
-	principals, _, err := c.MDSv2Client.RBACRoleBindingSummariesApi.LookupPrincipalsWithRole(
-		c.createContext(),
-		role,
-		*scopeV2)
-	if err != nil {
-		return err
+	var principals []string
+	var err error
+	if cmd.Flags().Changed("resource") {
+		r, err := cmd.Flags().GetString("resource")
+		if err != nil {
+			return err
+		}
+		resource, err := c.parseAndValidateResourcePatternV2(r, false)
+		if err != nil {
+			return err
+		}
+		err = c.validateRoleAndResourceTypeV2(role, resource.ResourceType)
+		if err != nil {
+			return err
+		}
+		principals, _, err = c.MDSv2Client.RBACRoleBindingSummariesApi.LookupPrincipalsWithRoleOnResource(
+			c.createContext(),
+			role,
+			resource.ResourceType,
+			resource.Name,
+			*scopeV2)
+		if err != nil {
+			return err
+		}
+	} else {
+		principals, _, err = c.MDSv2Client.RBACRoleBindingSummariesApi.LookupPrincipalsWithRole(
+			c.createContext(),
+			role,
+			*scopeV2)
+		if err != nil {
+			return err
+		}
 	}
 
 	userToEmailMap, err := c.userIdToEmailMap()
@@ -812,31 +882,52 @@ func (c *rolebindingCommand) parseCommon(cmd *cobra.Command) (*rolebindingOption
 	}
 
 	resourcesRequest := mds.ResourcesRequest{}
+	resourcesRequestV2 := mdsv2alpha1.ResourcesRequest{}
 	if resource != "" {
-		parsedResourcePattern, err := c.parseAndValidateResourcePattern(resource, prefix)
-		if err != nil {
-			return nil, err
-		}
-		// Resource types are defined under roles' access policies, so if no role is specified,
-		// we have to loop over the possible resource types for all roles (this is what
-		// validateResourceType does).
-		if role != "" {
-			err = c.validateRoleAndResourceType(role, parsedResourcePattern.ResourceType)
+		if c.cliName != "ccloud" {
+			parsedResourcePattern, err := c.parseAndValidateResourcePattern(resource, prefix)
 			if err != nil {
 				return nil, err
+			}
+
+			// Resource types are defined under roles' access policies, so if no role is specified,
+			// we have to loop over the possible resource types for all roles (this is what
+			// validateResourceType does).
+			if role != "" {
+				err = c.validateRoleAndResourceTypeV1(role, parsedResourcePattern.ResourceType)
+				if err != nil {
+					return nil, err
+				}
+			} else {
+				err = c.validateResourceType(parsedResourcePattern.ResourceType)
+				if err != nil {
+					return nil, err
+				}
+			}
+
+			resourcePatterns := []mds.ResourcePattern{
+				parsedResourcePattern,
+			}
+			resourcesRequest = mds.ResourcesRequest{
+				Scope:            *scope,
+				ResourcePatterns: resourcePatterns,
 			}
 		} else {
-			err = c.validateResourceType(parsedResourcePattern.ResourceType)
+			parsedResourcePattern, err := c.parseAndValidateResourcePatternV2(resource, prefix)
 			if err != nil {
 				return nil, err
 			}
-		}
-		resourcePatterns := []mds.ResourcePattern{
-			parsedResourcePattern,
-		}
-		resourcesRequest = mds.ResourcesRequest{
-			Scope:            *scope,
-			ResourcePatterns: resourcePatterns,
+			err = c.validateRoleAndResourceTypeV2(role, parsedResourcePattern.ResourceType)
+			if err != nil {
+				return nil, err
+			}
+			resourcePatterns := []mdsv2alpha1.ResourcePattern{
+				parsedResourcePattern,
+			}
+			resourcesRequestV2 = mdsv2alpha1.ResourcesRequest{
+				Scope:            *scopeV2,
+				ResourcePatterns: resourcePatterns,
+			}
 		}
 	}
 	return &rolebindingOptions{
@@ -847,6 +938,7 @@ func (c *rolebindingCommand) parseCommon(cmd *cobra.Command) (*rolebindingOption
 			*scopeV2,
 			*scope,
 			resourcesRequest,
+			resourcesRequestV2,
 		},
 		nil
 }
@@ -869,11 +961,19 @@ func (c *rolebindingCommand) confluentCreate(options *rolebindingOptions) (resp 
 }
 
 func (c *rolebindingCommand) ccloudCreate(options *rolebindingOptions) (*http.Response, error) {
-	return c.MDSv2Client.RBACRoleBindingCRUDApi.AddRoleForPrincipal(
-		c.createContext(),
-		options.principal,
-		options.role,
-		options.scopeV2)
+	if options.resource != "" {
+		return c.MDSv2Client.RBACRoleBindingCRUDApi.AddRoleResourcesForPrincipal(
+			c.createContext(),
+			options.principal,
+			options.role,
+			options.resourcesRequestV2)
+	} else {
+		return c.MDSv2Client.RBACRoleBindingCRUDApi.AddRoleForPrincipal(
+			c.createContext(),
+			options.principal,
+			options.role,
+			options.scopeV2)
+	}
 }
 
 func (c *rolebindingCommand) create(cmd *cobra.Command, _ []string) error {
@@ -905,18 +1005,37 @@ func (c *rolebindingCommand) create(cmd *cobra.Command, _ []string) error {
 
 func (c *rolebindingCommand) displayCCloudCreateAndDeleteOutput(cmd *cobra.Command, options *rolebindingOptions) error {
 	var fieldsSelected []string
-	structuredRename := map[string]string{"Principal": "principal", "Email": "email", "Role": "role"}
+	structuredRename := map[string]string{"Principal": "principal", "Email": "email", "Role": "role", "ResourceType": "resource_type", "Name": "name", "PatternType": "pattern_type"}
 	userResourceId := strings.TrimLeft(options.principal, "User:")
 	user, err := c.Client.User.Describe(context.Background(), &orgv1.User{ResourceId: userResourceId})
 	displayStruct := &listDisplay{
 		Principal: options.principal,
 		Role:      options.role,
 	}
+
+	if options.resource != "" {
+		if len(options.resourcesRequest.ResourcePatterns) != 1 {
+			return errors.New("display error: number of resource pattern is not 1")
+		}
+		resourcePattern := options.resourcesRequest.ResourcePatterns[0]
+		displayStruct.ResourceType = resourcePattern.ResourceType
+		displayStruct.Name = resourcePattern.Name
+		displayStruct.PatternType = resourcePattern.PatternType
+	}
+
 	if err != nil {
-		fieldsSelected = []string{"Principal", "Role"}
+		if options.resource != "" {
+			fieldsSelected = resourcePatternListFields
+		} else {
+			fieldsSelected = []string{"Principal", "Role"}
+		}
 	} else {
-		displayStruct.Email = user.Email
-		fieldsSelected = []string{"Principal", "Email", "Role"}
+		if options.resource != "" {
+			fieldsSelected = ccloudResourcePatternListFields
+		} else {
+			displayStruct.Email = user.Email
+			fieldsSelected = []string{"Principal", "Email", "Role"}
+		}
 	}
 	return output.DescribeObject(cmd, displayStruct, fieldsSelected, map[string]string{}, structuredRename)
 }
@@ -962,11 +1081,19 @@ func (c *rolebindingCommand) confluentDelete(options *rolebindingOptions) (resp 
 }
 
 func (c *rolebindingCommand) ccloudDelete(options *rolebindingOptions) (*http.Response, error) {
-	return c.MDSv2Client.RBACRoleBindingCRUDApi.DeleteRoleForPrincipal(
-		c.createContext(),
-		options.principal,
-		options.role,
-		options.scopeV2)
+	if options.resource != "" {
+		return c.MDSv2Client.RBACRoleBindingCRUDApi.RemoveRoleResourcesForPrincipal (
+			c.createContext(),
+			options.principal,
+			options.role,
+			options.resourcesRequestV2)
+	} else {
+		return c.MDSv2Client.RBACRoleBindingCRUDApi.DeleteRoleForPrincipal(
+			c.createContext(),
+			options.principal,
+			options.role,
+			options.scopeV2)
+	}
 }
 
 func (c *rolebindingCommand) delete(cmd *cobra.Command, _ []string) error {
